Done orders

+++ src/main/java/edu/unapec/shoppingorders/repositories/ShoppingOrderRepository.java
@@ -2,6 +2,6 @@ package edu.unapec.shoppingorders.repositories;
 
 import edu.unapec.shoppingorders.models.ShoppingOrder;
 
-public interface ShoppingOrderRepository extends Repository<ShoppingOrder, Integer> {
+public interface ShoppingOrderRepository extends Repository<ShoppingOrder, Long> {
 
 }
